Add in-app sync UI
The import script was useful for development, but SQLite means if the
app is running, we can't run a separate script and write to the same DB,
which kind of ruins it when it's on a server (can't cron job).

So, this moves the (luckily, now very simple) import code into a new
`POST /refresh` endpoint, which we add a button in the header to hit.